backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
firebase-admin
pydantic
httpx
openai
python-multipart
